backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
pre-commit
django-cors-headers
